+++ Libraries/RobsJuceModules/rapt/Spectral/HarmonicAnalyzer.cpp
@@ -274,7 +274,7 @@ void rsHarmonicAnalyzer<T>::handleEdges(RAPT::rsSinusoidalModel<T>& mdl)
   //T endTime = rsLast(tOut);
   T endTime = rsLast(tIn);
   rsInstantaneousSineParams<T> params;
-  for(k = 0; k < mdl.getNumPartials(); k++) 
+  for(k = 0; k < (int) mdl.getNumPartials(); k++) 
   {
     // fill first datapoint:
     params  = mdl.getPartial(k).getDataPoint(1);  
@@ -312,8 +312,8 @@ void rsHarmonicAnalyzer<T>::handleEdges(RAPT::rsSinusoidalModel<T>& mdl)
 template<class T>
 void rsHarmonicAnalyzer<T>::convertTimeUnit(RAPT::rsSinusoidalModel<T>& mdl)
 {
-  for(int hi = 0; hi < mdl.getNumPartials(); hi++)
-    for(int di = 0; di < getNumDataPoints(); di++)
+  for(size_t hi = 0; hi < mdl.getNumPartials(); hi++)
+    for(size_t di = 0; di < getNumDataPoints(); di++)
       mdl.getDataRef(hi, di).time /= sampleRate;
 }
