+++ Libraries/RobsJuceModules/rs_testing/Prototypes/Polygon.cpp
@@ -124,9 +124,9 @@ std::vector<rsVector2DF> clipAgainstEdge(const std::vector<rsVector2DF>& p,
   std::vector<rsVector2DF> r;
   if(p.size() == 0)
     return r;
-  Vec2 S = p[p.size()-1];               // start of edge under consideration
-  for(int i = 0; i < p.size(); i++) {   // loop over edges of polynomial
-    Vec2 E = p[i];                      // end of edge under consideration
+  Vec2 S = p[p.size()-1];                // start of edge under consideration
+  for(size_t i = 0; i < p.size(); i++) { // loop over edges of polynomial
+    Vec2 E = p[i];                       // end of edge under consideration
     if(isInsideEdge(E, e0, e1)) {
       if(!isInsideEdge(S, e0, e1))
         r.push_back(lineIntersection(S, E, e0, e1));
@@ -146,7 +146,7 @@ std::vector<rsVector2DF> clipPolygon(const std::vector<rsVector2DF>& p,
 {
   std::vector<rsVector2DF> r = p;
   Vec2 e0 = rsLast(c), e1;
-  for(int i = 0; i < c.size(); i++)
+  for(size_t i = 0; i < c.size(); i++)
   {
     e1 = c[i];
     r  = clipAgainstEdge(r, e0, e1);
